% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fills.R
\name{fills}
\alias{fills}
\title{Get List of Most Recent Fills}
\usage{
fills(api.key, secret, passphrase)
}
\arguments{
\item{api.key}{Mandatory character value. This is the API key as generated by GDAX. Typically a 32 character value.}

\item{secret}{Mandatory character value. This is the API secret as generated by GDAX. Typically a 88 character value.}

\item{passphrase}{THE URL component for the API. Default to "/accounts". Typically a 11 character value.}
}
\value{
Dataframe with fills for all products or for the provided products. The volume is quoted in USD.
}
\description{
Get a list of all previouslly filled orders.
}
\examples{
\dontrun{
fills(api.key = your_key, secret = your_api_secret, passphrase = your_api_pass)
}

}
