% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_orders.R
\name{open_orders}
\alias{open_orders}
\title{Get List of All Orders the User}
\usage{
open_orders(api.key, secret, passphrase)
}
\arguments{
\item{api.key}{Mandatory character value. This is the API key as generated by GDAX. Typically a 32 character value.}

\item{secret}{Mandatory character value. This is the API secret as generated by GDAX. Typically a 88 character value.}

\item{passphrase}{Mandatory character value. This is the passphrase as generated by GDAX. Typically a 11 character value.}
}
\value{
Dataframe with all orders and their status for that currency.
}
\description{
This is an auth based function. User must have valid api keys generated by GDAX which must be passed as mandatory arguments. The function will return all open orders for all currencies. This is an extendion of \code{\link{holds}} & \code{\link{fills}}.
}
\examples{
\dontrun{
order_info(api.key = your_key, secret = your_api_secret, passphrase = your_api_pass)
}

}
