% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_daystats.R
\name{public_daystats}
\alias{public_daystats}
\title{24 Hour Stats For A Given Product}
\usage{
public_daystats(product_id = "LTC-USD")
}
\arguments{
\item{product_id}{Optional character parameter. This is a case insensitive value of the product id for which the order book is desired. Default to \code{'LTC-USD'}. For all valid product ids, refer to \code{\link{public_info}}.}
}
\value{
Dataframe with a single row of product's last 24 hour stats.
}
\description{
This function is a public function and will fetch all bids/asks for provided currency pair (product_id). User should change the level to get the best bid/ask or to fetch all bids/asks.
}
\examples{
\dontrun{
public_daystats()
}

}
