% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_info.R
\name{public_info}
\alias{public_info}
\title{Get all supported products (currency - pairs) by GDAX.}
\usage{
public_info(product = TRUE)
}
\arguments{
\item{product}{Optional Boolean Parameter. Default to \code{TRUE}. \code{NA} behavior is similar to default. \code{FALSE} is equivalent of currency public end point.}
}
\value{
Dataframe with ALL supported currency pairs.  A dataframe of base currencies is returned when the flag is et as \code{FALSE}.
}
\description{
This function is a public function and will fetch all supported currency pairs by default. The function can also fetch all the supported currencies based on the \code{source} argument.
}
\examples{
\dontrun{
public_info()
public_info(product = FALSE)
}

}
