% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_time.R
\name{public_time}
\alias{public_time}
\title{Get GDAX API Server Time}
\usage{
public_time()
}
\value{
Dataframe with time in ISO and the epoch field represents decimal seconds since Unix Epoch.
}
\description{
Gets the server time from GDAX for reference purposes. This function does not take any arguments.
}
\examples{
\dontrun{
public_time()
}

}
