% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pymt_methods.R
\name{pymt_methods}
\alias{pymt_methods}
\title{View All Linked payment Methods.}
\usage{
pymt_methods(api.key, secret, passphrase)
}
\arguments{
\item{api.key}{Mandatory character value. This is the API key as generated by GDAX. Typically a 32 character value.}

\item{secret}{Mandatory character value. This is the API secret as generated by GDAX. Typically a 88 character value.}

\item{passphrase}{THE URL component for the API. Default to "/accounts". Typically a 11 character value.}
}
\value{
A named list.
}
\description{
TBD
}
\examples{
\dontrun{
pmt_methods(api.key = your_key, secret = your_api_secret, passphrase = your_api_pass)
}

}
