% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_response}
\alias{parse_response}
\title{Utility Function to parse message from api response}
\usage{
parse_response(path, query = NULL)
}
\arguments{
\item{path}{Mandatory character parameter. This is an extension of the api end point and passed as an argument from the function calling \code{parse_response}.}

\item{query}{Optional named list parameter. This would consist of any named params to be passed to the api end point.}
}
\value{
A named list or a dataframe object of the response.
}
\description{
An internal function to parse the api response for various public functions.
}
\examples{
\dontrun{
parse_response("/currencies")
parse_response(path = your_path, query = list("start" = start,"end" = end))
}

}
