% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_daystats.R
\name{public_daystats}
\alias{public_daystats}
\title{24 Hour Stats For A Given Product}
\usage{
public_daystats(product_id = "BTC-USD")
}
\arguments{
\item{product_id}{Optional character parameter. This is a case insensitive value of the product id for which the order book is desired. Default to \code{'BTC-USD'}. For all valid product ids, refer to \code{\link{public_info}}.}
}
\value{
Dataframe with a single row of product's last 24 hour stats.
}
\description{
This function is a public function and will fetch get 24 hr stats for the provided currency pair (product_id). Volume is in base currency units. open, high, low are in quote currency units.
}
\examples{
\dontrun{
public_daystats()
}

}
