% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pymt_methods.R
\name{pymt_methods}
\alias{pymt_methods}
\title{View All Linked payment Methods.}
\usage{
pymt_methods(api.key, secret, passphrase)
}
\arguments{
\item{api.key}{Mandatory character value. This is the API key as generated by GDAX. Typically a 32 character value.}

\item{secret}{Mandatory character value. This is the API secret as generated by GDAX. Typically a 88 character value.}

\item{passphrase}{Mandatory character value. This is the passphrase as generated by GDAX. Typically a 11 character value.}
}
\value{
A named list.
}
\description{
This is an auth based function. User must have valid api keys generated by GDAX which must be passed as mandatory arguments. These keys must have the appropriate access. Note that for this function to work, the API key must have transfer access.
}
\examples{
\dontrun{
pmt_methods(api.key = your_key, secret = your_api_secret, passphrase = your_api_pass)
}

}
